+++ src/ios/OSCAMRParametersStruct.swift
@@ -4,10 +4,12 @@ import OSCameraLib
 struct OSCAMRChooseGalleryParameters {
     let mediaType: OSCAMRMediaType
     let allowMultipleSelection: Bool
+    let includeMetadata: Bool
     
-    init(mediaType: OSCAMRMediaType, allowMultipleSelection: Bool) {
+    init(mediaType: OSCAMRMediaType, allowMultipleSelection: Bool, includeMetadata: Bool) {
         self.mediaType = mediaType
         self.allowMultipleSelection = allowMultipleSelection
+        self.includeMetadata = includeMetadata
     }
 }
 
@@ -15,15 +17,17 @@ extension OSCAMRChooseGalleryParameters: Decodable {
     enum CodingKeys: String, CodingKey {
         case mediaType
         case allowMultipleSelection
+        case includeMetadata
     }
     
     init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         let mediaTypeValue = try container.decode(Int.self, forKey: .mediaType)
         let allowMultipleSelection = try container.decode(Bool.self, forKey: .allowMultipleSelection)
+        let includeMetadata = try container.decode(Bool.self, forKey: .includeMetadata)
                 
         let mediaType = try OSCAMRMediaType(from: mediaTypeValue)
-        self.init(mediaType: mediaType, allowMultipleSelection: allowMultipleSelection)
+        self.init(mediaType: mediaType, allowMultipleSelection: allowMultipleSelection, includeMetadata: includeMetadata)
     }
 }
 
@@ -51,5 +55,37 @@ extension OSCAMRPlayVideoParameters: Decodable {
         
         guard let url = URL(string: urlString) else { throw DecodeError.invalidURL }
         self.init(url: url)
+    }
+}
+
+struct OSCAMRRecordVideoParameters {
+    let saveToGallery: Bool
+    let includeMetadata: Bool
+    
+    init(saveToGallery: Bool, includeMetadata: Bool) {
+        self.saveToGallery = saveToGallery
+        self.includeMetadata = includeMetadata
+    }
+}
+
+extension OSCAMRRecordVideoParameters: Decodable {
+    enum CodingKeys: String, CodingKey {
+        case saveToGallery
+        case includeMetadata
+    }
+    
+    init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        
+        let saveToGallery = try container.decode(Bool.self, forKey: .saveToGallery)
+        let includeMetadata = try container.decode(Bool.self, forKey: .includeMetadata)
+        
+        self.init(saveToGallery: saveToGallery, includeMetadata: includeMetadata)
+    }
+}
+
+extension OSCAMRVideoOptions {
+    convenience init(from parameters: OSCAMRRecordVideoParameters) {
+        self.init(saveToPhotoAlbum: parameters.saveToGallery, returnMetadata: parameters.includeMetadata)
     }
 }